requirements.txt
arabic-reshaper==3.0.0
Django==4.1.5
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.20.2
numpy==1.24.3
oauthlib==3.2.0
optional-django==0.1.0
SQLAlchemy==2.0.9
sqlparse==0.4.3
texttable==1.6.4
tokenizers==0.13.3
torch==2.0.1
transformers==4.29.2
scipy==1.10.1
python-bidi==0.4.2
